Make i-PI calculator into a generic ML-MD interface
Remove i-PI specific unit conversions and input/output formatting; this
should be taken care of on the i-PI side.
